+++ xpaw/downloader.py
@@ -61,8 +61,8 @@ class DownloaderMiddlewareManager(MiddlewareManager):
             self._error_handlers.insert(0, coro_wrapper(middleware.handle_error))
 
     @classmethod
-    def _middleware_list_from_config(cls, config):
-        mw_list = config.get("downloader_middlewares")
+    def _middleware_list_from_cluster(cls, cluster):
+        mw_list = cluster.config.get("downloader_middlewares")
         if mw_list:
             if not isinstance(mw_list, list):
                 mw_list = [mw_list]